[DRIVERS] Freescale i.MX eSDHC controller driver.
This patch adds the support for the Freescale i.MX eSDHC controller.

Note:
The original i.MX uSDHC driver expected the "esdhc_pltfm_bus_width" function to
be called, managing the bus width in the host controller register.
Since this function is called by the "platform_bus_width" operation, removed in
Xvisor, this must be done the usual way, hence when writing this register.