[CORE] Avoid reading the next pending IRQ in case of IRQ burst avoidance.
In the case of the ARMv7 GIC, reading the first pending IRQ set it as active,
disabling the others with a lower priority.
If we have an IRQ burst, we may stop the "vmm_host_active_irq_exec" loop, but
we must not read the pending IRQ before leaving, to avoid marking it as active.